test: split admin integration tests
