+++ crates/dcc-mcp-gateway/src/gateway/admin/mod.rs
@@ -79,6 +79,8 @@ mod analytics_tests;
 mod handlers;
 #[cfg(all(test, feature = "admin"))]
 mod instance_update_tests;
+#[cfg(all(test, feature = "admin"))]
+mod integration_tests;
 #[cfg(feature = "admin")]
 mod router;
 #[cfg(all(test, feature = "admin"))]
